requirements.txt
langchain
langchain-openai

gradio
python-dotenv

chromadb